pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kms-sfdc-vector-db"
version = "0.1.0"
description = "HPE KMS SFDC Vector Database for Case Similarity Search"
authors = [{name = "HPE Team"}]
license = {text = "Proprietary"}
requires-python = ">=3.8.1"
dependencies = [
    # Core dependencies
    "faiss-cpu==1.7.4",
    "sentence-transformers>=3.0.0",
    "numpy==1.24.3",
    "einops>=0.7.0",
    "pandas==2.0.3",
    "scikit-learn==1.3.0",
    # Salesforce API
    "simple-salesforce==1.12.4",
    "requests==2.31.0",
    # Data processing
    "nltk==3.8.1",
    "spacy==3.6.1",
    "textblob==0.17.1",
    # Configuration and logging
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    "loguru==0.7.0",
    # API framework
    "fastapi==0.103.1",
    "uvicorn==0.23.2",
    "pydantic==2.3.0",
    # Database and caching
    "redis==4.6.0",
    "pymongo==4.5.0",
    # Monitoring and performance
    "prometheus-client==0.17.1",
    "psutil==5.9.5",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.2",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.21.1",
    
    # Development tools
    "black==23.7.0",
    "flake8==6.0.0",
    "mypy==1.5.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.1",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow running tests"
]
